fix: tests

+++ src/stack.cairo
@@ -191,7 +191,7 @@ pub impl ScriptStackImpl of ScriptStackTrait {
         return Result::Ok(value);
     }
 
-    fn pick_n(ref self: ScriptStack, idx: i64) -> Result<(), felt252> {
+    fn pick_n(ref self: ScriptStack, idx: i32) -> Result<(), felt252> {
         let so = self.peek_byte_array(idx.try_into().unwrap())?;
         self.push_byte_array(so);
         return Result::Ok(());
